fix: separate datalad_integration functions from ONHarmony specific functions
