+++ src/uniharmony/datasets/__init__.pyi
@@ -1,7 +1,11 @@
 __all__ = [
+    "_check_datalad_installed",
+    "_ensure_hidden_dataset",
     "_get_onharmony_information_form_idps",
     "_list_available_files",
+    "_list_available_files",
     "_list_available_possibilities",
+    "_make_visible_directory",
     "get_multisite_data_statistics",
     "load_MAREoS",
     "load_onharmony",
@@ -9,10 +13,10 @@ __all__ = [
     "print_statistics_summary",
 ]
 
+from ._datalad_integration import _check_datalad_installed, _ensure_hidden_dataset, _list_available_files, _make_visible_directory
 from ._load_mareos import load_MAREoS
 from ._load_onharmony import (
     _get_onharmony_information_form_idps,
-    _list_available_files,
     _list_available_possibilities,
     load_onharmony,
 )
